requirements.txt
guided-filter-pytorch==3.7.5
numba==0.45.1
Pillow==6.1.0
scikit-image==0.15.0
torch==0.4.1
torchvision==0.2.0
